file ex17.py - upgrade

+++ ex17.py
@@ -10,15 +10,17 @@ in_file = open(from_file)
 indata = in_file.read()
 
 print(f"The input file is {len(indata)} bytes long")
-
 print(f"Does the output file exist? {exists(to_file)}")
-print("Ready, hit RETURN to continue, CTRL_C to abort.")
-input()
+if exists(to_file) == True:
+    print("Ready, hit RETURN to continue, CTRL_C to abort.")
+    input()
 
-out_file = open(to_file, 'w')
-out_file.write(indata)
+    out_file = open(to_file, 'w')
+    out_file.write(indata)
 
-print("Alright, all done.")
+    print("Alright, all done.")
 
-out_file.close()
-in_file.close()
+    out_file.close()
+    in_file.close()
+elif exists(to_file) == False:
+    print("Output file doesn't exist.")
